requirements.txt

streamlit
pandas
plotly
fpdf
google-generativeai
google-genai
yfinance
astral
